refactor(fresco): split raw-mode platform backends into modules
